Tests: Move helper functions to tests/helper.py
- move helper functions to their own module in helper.py so they can be
  used by multiple test modules
